add refactors3

+++ net-http/pkg/core/core.go
@@ -24,19 +24,19 @@ func ExtractFieldsFromStruct[T any](t *T) []any {
 	return f
 }
 
-func CastTypeIfInt(x any) int {
+func CastTypeIfInt(x any) int32 {
 	switch v := x.(type) {
 	case int, int32, int64 :
-		return v.(int)
+		return v.(int32)
 	default:
 		return -1
 	}
 }
 
-func CastTypeIfFloat(x any) float64 {
+func CastTypeIfFloat(x any) float32 {
 	switch v := x.(type) {
 	case float32, float64 :
-		return v.(float64)
+		return v.(float32)
 	default:
 		return -1
 	}
